Refactor core library config to not be limited to only particle modules
